+++ CHANGELOG.md
@@ -18,7 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Removed `faulty` producer assertion in integration tests for better performance
 * Integration tests use XUnit for parallelization. Updates `TestFixture`, `TestBase`, and test attributes
   * Note, there is still no db cleanup after each test, so you should write your assertions accordingly
-* Unit tests use XUnit for parallelization.
+* Unit tests use XUnit for parallelization
+* Integration tests updated to use `DotNet.Testcontainers` for simpler docker db setup
 
 ## [0.15.0] - 06/05/2022
 
@@ -56,7 +57,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   }
   ```
   
-
 * Smart enum property support. Just add a property and give it a list of options using the `SmartNames` property and you're good to go. They will all be strings with smart enum backing. Attributes like filtering and sorting are supported
 
   ```yaml